+++ easier/core/dump.py
@@ -24,7 +24,7 @@ from easier.core.passes.utils import EasierInterpreter, OrderedSet, \
     fx_graph_to_serializable_ir, serializable_ir_to_fx_graph
 from easier.core.runtime.dist_env import get_runtime_dist_env
 from easier.core.runtime.modules import HaloExchanger
-from easier.core.runtime.data_loader import \
+from easier.core.runtime.data_loader.data_loader import \
     DataLoaderBase, InMemoryTensorLoader
 from easier.core.passes.tensor_group_partition import \
     ElemPart, ElemPartArangeIdx
