+++ easier/core/runtime/data_loader/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+
+from easier.core.runtime.data_loader.base import (
+    H5DataLoader
+)
